refactor(guest-agent): centralize codex startup policy fixture
