+++ crates/guest-agent/tests/codex_app_server_backend.rs
@@ -3,22 +3,14 @@
 //! This test lives in its own binary to isolate process env, working directory,
 //! and guest runtime path overrides used during setup.
 
+#[path = "common/codex_app_server_startup_policy.rs"]
+mod codex_app_server_startup_policy;
 mod common;
 
 use guest_agent::masker::SecretMasker;
 use serde_json::Value;
-use shell_quote::quote_shell_arg;
 use std::collections::HashMap;
-use std::path::{Path, PathBuf};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
-const CODEX_FIXED_STARTUP_CONFIGS: [&str; 5] = [
-    "analytics.enabled=false",
-    "features.plugins=false",
-    "features.apps=false",
-    "features.goals=false",
-    "features.image_generation=false",
-];
 
 #[tokio::test]
 async fn codex_app_server_backend_runs_initial_turn_and_synthesizes_thread_started()
@@ -26,7 +18,8 @@ async fn codex_app_server_backend_runs_initial_turn_and_synthesizes_thread_start
     let mock = common::build_and_locate_mock_codex()?;
     let tmp = tempfile::tempdir()?;
     let argv_path = tmp.path().join("codex-argv");
-    let recording_mock = recording_codex(tmp.path(), &mock, &argv_path)?;
+    let recording_mock =
+        codex_app_server_startup_policy::recording_codex(tmp.path(), &mock, &argv_path)?;
     let run_id = "codex-app-server-backend-test";
     let prompt = "drive the app-server backend";
 
@@ -93,7 +86,7 @@ async fn codex_app_server_backend_runs_initial_turn_and_synthesizes_thread_start
     assert_eq!(cli_result.exit_code, common::CLEAN_EXIT);
     assert!(cli_result.failure_diagnostic.is_none());
     assert!(cli_result.last_event_sequence.is_none());
-    assert_app_server_fixed_startup_policy(&argv_path)?;
+    codex_app_server_startup_policy::assert_startup_policy(&argv_path, false)?;
 
     let events = read_agent_log_events(&runtime.paths)?;
     assert_event_type_sequence(
@@ -283,47 +276,4 @@ fn assert_event_type_sequence(events: &[Value], expected: &[&str]) {
         .map(|value| Some(*value))
         .collect::<Vec<_>>();
     assert_eq!(actual, expected);
-}
-
-fn recording_codex(root: &Path, mock: &Path, argv_path: &Path) -> Result<PathBuf, std::io::Error> {
-    use std::os::unix::fs::PermissionsExt as _;
-
-    let path = root.join("recording-codex");
-    std::fs::write(
-        &path,
-        format!(
-            "#!/bin/sh\nprintf '%s\\n' \"$@\" > {}\nexec {} \"$@\"\n",
-            quote_shell_arg(&argv_path.to_string_lossy()),
-            quote_shell_arg(&mock.to_string_lossy()),
-        ),
-    )?;
-    let mut permissions = std::fs::metadata(&path)?.permissions();
-    permissions.set_mode(0o700);
-    std::fs::set_permissions(&path, permissions)?;
-    Ok(path)
-}
-
-fn assert_app_server_fixed_startup_policy(argv_path: &Path) -> Result<(), std::io::Error> {
-    let args = std::fs::read_to_string(argv_path)?
-        .lines()
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    let app_server_index = args
-        .iter()
-        .position(|arg| arg == "app-server")
-        .ok_or_else(|| std::io::Error::other("Codex argv omitted app-server subcommand"))?;
-    for expected in CODEX_FIXED_STARTUP_CONFIGS {
-        let config_index = args
-            .windows(2)
-            .position(|window| {
-                matches!(window, [flag, value] if flag == "-c" && value == expected)
-            })
-            .ok_or_else(|| {
-                std::io::Error::other(format!(
-                    "Codex app-server argv should include fixed startup config {expected:?}: {args:?}"
-                ))
-            })?;
-        assert!(config_index < app_server_index);
-    }
-    Ok(())
 }
